fix: refactor page index structure position
